requirements.txt
transformers>=4.49.0
tiktoken>=0.9.0
seaborn>=0.13.2
matplotlib>=3.10.1
sentence-transformers>=4.0.2
pandas>=2.2.3
openai>=1.68.2
google-genai>=1.9.0 
jieba>=0.42.1
nltk>=3.9.1
tqdm>=4.67.1